Work for Exercise 5 - Natalie
I added my work for exercise 5 to the bottom of the cdc exercise qmd file that you created. For the most part, I was able to successfully create a synthetic dataset. However, I was unable to replicate the last two graphs.
For some reason, GitKraken was showing that the NFL draft picture was going to be deleted again - so, I just unstaged that change (but kept all other changes). I hope that this works! :)